+++ src/app/ilm.component.ts
@@ -13,6 +13,9 @@ import { Observable } from 'rxjs/Observable';
 import { TaxaDatabase } from './db/taxa.database';
 import { LoginComponent } from './login/login.component';
 import { Stored, StoreService } from './services/store.service';
+import { DialogsService } from './services/dialog.service';
+import { WindowRef } from './ref/window.ref';
+import { setTimeout } from 'timers';
 
 @Component({
   selector: 'ilm-root',
@@ -34,11 +37,13 @@ export class IlmComponent implements OnInit {
   constructor(
     sanitizer: DomSanitizer,
     iconRegistry: MdIconRegistry,
+    private windowRef: WindowRef,
     private docDB: DocumentDatabase,
     private storeDB: StoreDatabase,
     private taxaDB: TaxaDatabase,
     private storeService: StoreService,
     private locationService: LocationStoreService,
+    private dialogSerivce: DialogsService,
     public snackBar: MdSnackBar,
     public formService: FormService,
     public userService: UserService,
@@ -123,12 +128,18 @@ export class IlmComponent implements OnInit {
   }
 
   clearLocal() {
-    this.docDB.destroy();
-    this.storeDB.destroy();
-    this.taxaDB.destroy();
-    this.snackBar.open('Tiedot poistettu', undefined, {
-      duration: 1500
-    });
+    this.dialogSerivce.confirm('Oletko varma että', 'haluat tyhjentään kaikki paikallisesti tallennetun datan?')
+      .subscribe((confirm) =>  {
+        if (confirm) {
+          this.docDB.destroy();
+          this.storeDB.destroy();
+          this.taxaDB.destroy();
+          this.snackBar.open('Tiedot poistettu', undefined, {
+            duration: 1500
+          });
+          setTimeout(() => this.windowRef.nativeWindow.location.reload(), 2000);
+        }
+      });
   }
 
   openInfo(type, options = {}): MdDialogRef<InfoComponent> {
